Moved webcast swapping to a standalone dialog that matches the aspect ratio of the window
