+++ react/gameday2/components/WebcastSelectionDialog.js
@@ -1,6 +1,5 @@
 import React, { PropTypes } from 'react'
 import Dialog from 'material-ui/Dialog'
-import Paper from 'material-ui/Paper'
 import { green500, indigo500 } from 'material-ui/styles/colors'
 import Divider from 'material-ui/Divider'
 import FlatButton from 'material-ui/FlatButton'
@@ -10,7 +9,6 @@ import ActionGrade from 'material-ui/svg-icons/action/grade'
 import ActionHelp from 'material-ui/svg-icons/action/help'
 import VideoCam from 'material-ui/svg-icons/av/videocam'
 import VideoCamOff from 'material-ui/svg-icons/av/videocam-off'
-import EventListener from 'react-event-listener'
 import WebcastSelectionDialogItem from './WebcastSelectionDialogItem'
 import { webcastPropType } from '../utils/webcastUtils'
